+++ src/table/handlers/TableHandler.ts
@@ -188,6 +188,8 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
       throw StorageErrorFactory.getPropertiesNeedValue(context);
     }
 
+    this.checkProperties(context, options.tableEntityProperties);
+
     const entity: Entity = {
       PartitionKey: options.tableEntityProperties.PartitionKey,
       RowKey: options.tableEntityProperties.RowKey,
@@ -273,13 +275,13 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
   private static getAndCheck(
     key: string | undefined,
     getFromContext: () => string,
-    contextForThrow: Context,
+    contextForThrow: Context
   ): string {
     if (key !== undefined) {
       return key;
     }
 
-    const fromContext = getFromContext()
+    const fromContext = getFromContext();
     if (fromContext === undefined) {
       throw StorageErrorFactory.getPropertiesNeedValue(contextForThrow);
     }
@@ -291,12 +293,20 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     partitionKey: string | undefined,
     rowKey: string | undefined,
     tableContext: TableStorageContext,
-    contextForThrow: Context,
+    contextForThrow: Context
   ) {
-    partitionKey = TableHandler.getAndCheck(partitionKey, () => tableContext.partitionKey!, contextForThrow);
-    rowKey = TableHandler.getAndCheck(rowKey, () => tableContext.rowKey!, contextForThrow);
+    partitionKey = TableHandler.getAndCheck(
+      partitionKey,
+      () => tableContext.partitionKey!,
+      contextForThrow
+    );
+    rowKey = TableHandler.getAndCheck(
+      rowKey,
+      () => tableContext.rowKey!,
+      contextForThrow
+    );
 
-    return [partitionKey, rowKey]
+    return [partitionKey, rowKey];
   }
 
   // TODO: Create data structures to hold entity properties and support serialize, merge, deserialize, filter
@@ -313,7 +323,12 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     const account = this.getAndCheckAccountName(tableContext);
     const table = this.getAndCheckTableName(tableContext);
 
-    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(partitionKey, rowKey, tableContext, context);
+    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(
+      partitionKey,
+      rowKey,
+      tableContext,
+      context
+    );
 
     const ifMatch = options.ifMatch;
 
@@ -329,6 +344,7 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
         `TableHandler:updateEntity() Incoming PartitionKey:${partitionKey} RowKey:${rowKey} in URL parameters don't align with entity body PartitionKey:${options.tableEntityProperties.PartitionKey} RowKey:${options.tableEntityProperties.RowKey}.`
       );
     }
+    this.checkProperties(context, options.tableEntityProperties);
 
     // Test if etag is available
     // this is considered an upsert if no etag header, an empty header is an error.
@@ -398,7 +414,12 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     const account = this.getAndCheckAccountName(tableContext);
     const table = this.getAndCheckTableName(tableContext);
 
-    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(partitionKey, rowKey, tableContext, context);
+    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(
+      partitionKey,
+      rowKey,
+      tableContext,
+      context
+    );
 
     if (!options.tableEntityProperties) {
       throw StorageErrorFactory.getPropertiesNeedValue(context);
@@ -417,7 +438,7 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
         `TableHandler:mergeEntity() Incoming PartitionKey:${partitionKey} RowKey:${rowKey} in URL parameters don't align with entity body PartitionKey:${options.tableEntityProperties.PartitionKey} RowKey:${options.tableEntityProperties.RowKey}.`
       );
     }
-
+    this.checkProperties(context, options.tableEntityProperties);
     const eTag = newTableEntityEtag(context.startTime!);
 
     const entity: Entity = {
@@ -472,7 +493,12 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     const tableContext = new TableStorageContext(context);
     const accountName = tableContext.account;
 
-    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(partitionKey, rowKey, tableContext, context);
+    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(
+      partitionKey,
+      rowKey,
+      tableContext,
+      context
+    );
 
     if (ifMatch === "" || ifMatch === undefined) {
       throw StorageErrorFactory.getPreconditionFailed(context);
@@ -612,7 +638,12 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     const account = this.getAndCheckAccountName(tableContext);
     const table = _table ? _table : this.getAndCheckTableName(tableContext);
 
-    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(partitionKey, rowKey, tableContext, context);
+    [partitionKey, rowKey] = TableHandler.getAndCheckKeys(
+      partitionKey,
+      rowKey,
+      tableContext,
+      context
+    );
 
     const accept = this.getAndCheckPayloadFormat(tableContext);
 
@@ -931,5 +962,34 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
       response.preferenceApplied = RETURN_CONTENT;
     }
     return response;
+  }
+
+  /**
+   * Checks that properties are valid according to rules given here:
+   * https://docs.microsoft.com/en-us/rest/api/storageservices/understanding-the-table-service-data-model#property-types
+   *
+   * @private
+   * @param {Context} context
+   * @param {{
+   *       [propertyName: string]: any;
+   *     }} properties
+   * @memberof TableHandler
+   */
+  private checkProperties(
+    context: Context,
+    properties: {
+      [propertyName: string]: any;
+    }
+  ) {
+    for (const prop in properties) {
+      if (properties.hasOwnProperty(prop)) {
+        if (
+          undefined !== properties[prop].length &&
+          properties[prop].length > 1024 * 32
+        ) {
+          throw StorageErrorFactory.getPropertyValueTooLargeError(context);
+        }
+      }
+    }
   }
 }